fixing issue with urwell protype dimensions not being set before creation of geant4 volumes
