+++ common-tools/clas-jcsg/src/main/java/org/jlab/detector/geant4/v2/MPGD/MUVT/MUVTStripFactory.java
@@ -24,8 +24,9 @@ public final class MUVTStripFactory extends MPGDTrapezoidStripFactory {
 
     /**
      * Build using an already-configured DatabaseConstantProvider.
+     * @param cp
      */
-    public MUVTStripFactory(DatabaseConstantProvider cp, String variation) {
+    public MUVTStripFactory(DatabaseConstantProvider cp) {
         super(new MUVTConstants(cp));
