+++ common-tools/clas-jcsg/src/main/java/org/jlab/detector/geant4/v2/MPGD/trapezoid/MPGDTrapezoidConstants.java
@@ -1,5 +1,6 @@
 package org.jlab.detector.geant4.v2.MPGD.trapezoid;
 
+import eu.mihosoft.vrl.v3d.Vector3d;
 import org.jlab.detector.calib.utils.DatabaseConstantProvider;
 
 import java.util.LinkedHashMap;
@@ -69,6 +70,17 @@ public class MPGDTrapezoidConstants {
     public static final double XENLARGEMENT = 0.15; // cm
     public static final double YENLARGEMENT = 0.15; // cm
     public static final double ZENLARGEMENT = 0.02; // cm
+
+
+
+    public static record SectorDimensions(
+            double halfThickness,
+            double halfHeight,
+            double halfLargeBase,
+            double halfSmallBase,
+            double tiltRad
+            ) {
+    }
 
     // ------------------------------------------------------------------------
     //  Material description: (layer, component) -> parameters
@@ -221,4 +233,148 @@ public class MPGDTrapezoidConstants {
                     .put(component, info);
         }
     }
+    
+    // ------------------------------------------------------------------------
+    //  Sector geometry helpers
+    // ------------------------------------------------------------------------
+    /**
+     * Computes the total thickness (mm) of a sector by summing the thickness of
+     * all material volumes.
+     *
+     * @return
+     */
+    public double getSectorThickness() {
+        return detectorStructure.values()
+                .stream()
+                .flatMap(componentMap -> componentMap.values().stream())
+                .mapToDouble(info -> info.thickness)
+                .sum();
+    }
+
+    /**
+     *
+     * @param region
+     * @return
+     */
+    public SectorDimensions getSectorActiveVolumeDimensions(int region) {
+
+        double baseDistance = this.TGTDET + region * this.DZ;
+
+        double sectorHeight = baseDistance
+                * (Math.tan(Math.toRadians(this.THMAX - this.THTILT))
+                + Math.tan(Math.toRadians(this.THTILT - this.THMIN)));
+
+        double halfThickness = this.getSectorThickness() / 2.0;
+        double halfHeight = sectorHeight / 2.0;
+
+        // Distance from target to the bottom base along the tilted axis
+        double W2TGT = (this.TGTDET + region * this.DZ)
+                / Math.cos(Math.toRadians(this.THTILT - this.THMIN));
+
+        double YMIN = W2TGT * Math.sin(Math.toRadians(this.THMIN)); // distance from beamline (Y)        
+        double h = sectorHeight * Math.cos(Math.toRadians(this.THTILT));
+        double halfSmallBase = 0.5 * (YMIN * Math.tan(Math.toRadians(this.THOPEN) / 2));
+
+        double halfLargeBase = halfSmallBase + sectorHeight * Math.tan(Math.toRadians(this.THOPEN / 2.0));
+
+        double tiltRad = Math.toRadians(this.THTILT);
+
+        double twidth_Check = 2 * halfLargeBase * Math.sin(Math.toRadians(this.THOPEN));
+
+        if (MPGDTrapezoidConstants.VERBOSE) {
+            System.out.printf("this.TWIDT=%.3f vs %.3f", this.TWIDTH, twidth_Check);
+
+            System.out.printf("YMIN=%.3f", YMIN);
+
+            System.out.printf(
+                    "SectorDimensionsPhysical [%s] region=%d : height=%.3f | halfT=%.3f halfH=%.3f "
+                    + "halfLarge=%.3f halfSmall=%.3f tilt(deg)=%.3f%n",
+                    this.detectorName, region, sectorHeight,
+                    halfThickness, halfHeight,
+                    halfLargeBase, halfSmallBase,
+                    this.THTILT
+            );
+        }
+
+        return new SectorDimensions(halfThickness, halfHeight, halfLargeBase, halfSmallBase, tiltRad);
+    }
+
+    /**
+     *
+     * @param region
+     * @return
+     */
+    public SectorDimensions getSectorContainerDimensions(int region) {
+
+        SectorDimensions phys = getSectorActiveVolumeDimensions(region);
+
+        double halfThickness = phys.halfThickness() + MPGDTrapezoidConstants.ZENLARGEMENT;
+        double halfHeight = phys.halfHeight() + MPGDTrapezoidConstants.YENLARGEMENT;
+        double halfLargeBase = phys.halfLargeBase() + MPGDTrapezoidConstants.XENLARGEMENT;
+        double halfSmallBase = phys.halfSmallBase() + MPGDTrapezoidConstants.XENLARGEMENT;
+
+        return new SectorDimensions(halfThickness, halfHeight, halfLargeBase, halfSmallBase, phys.tiltRad());
+    }
+
+    /**
+     * Computes the sector height (longitudinal extension in the RZ plane) for a
+     * given region.
+     *
+     * @param region
+     * @return
+     */
+    public double getSectorHeight(int region) {
+
+        double baseDistance = this.TGTDET + region * this.DZ;
+
+        double sectorHeight = baseDistance
+                * (Math.tan(Math.toRadians(this.THMAX - this.THTILT))
+                + Math.tan(Math.toRadians(this.THTILT - this.THMIN)));
+
+        if (MPGDTrapezoidConstants.VERBOSE) {
+            System.out.printf(
+                    "SectorHeight [%s] region=%d : baseDistance=%.3f THMIN=%.3f THMAX=%.3f THTILT=%.3f -> height=%.3f%n",
+                    this.detectorName,
+                    region,
+                    baseDistance,
+                    this.THMIN, this.THMAX, this.THTILT,
+                    sectorHeight
+            );
+        }
+
+        return sectorHeight;
+    }
+
+    /**
+     * Computes the barycenter coordinates of a given sector/region in the
+     * CLAS12 coordinate system.
+     *
+     * @param isector
+     * @param iregion
+     * @return
+     */
+    public Vector3d getCenterCoordinate(int isector, int iregion) {
+
+        Vector3d vCenter = new Vector3d(0, 0, 0);
+
+        // Distance from target to the bottom base along the tilted axis
+        double W2TGT = (this.TGTDET + iregion * this.DZ)
+                / Math.cos(Math.toRadians(this.THTILT - this.THMIN));
+
+        double YMIN = W2TGT * Math.sin(Math.toRadians(this.THMIN)); // distance from beamline (Y)
+        double ZMIN = W2TGT * Math.cos(Math.toRadians(this.THMIN)); // Z of the bottom base
+
+        SectorDimensions dimCont = this.getSectorContainerDimensions(iregion);
+        double sectorHeight = 2 * dimCont.halfHeight();
+
+        vCenter.x = 0.0;
+        vCenter.y = (sectorHeight / 2.0) * Math.cos(Math.toRadians(this.THTILT)) + YMIN;
+        vCenter.z = -(sectorHeight / 2.0) * Math.sin(Math.toRadians(this.THTILT)) + ZMIN;
+
+        // Rotate to the correct sector around Z (assumes 6 sectors, 60° apart)
+        vCenter.rotateZ(-Math.toRadians(90.0 - isector * 60.0));
+
+        return vCenter;
+    }
+    
 }
